backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
qdrant-client>=1.8.0
google-generativeai>=0.8.0
groq>=0.4.0
python-dotenv>=1.0.0